Embed the jvm
For this to work copy the contents of `java-14-openjdk` into `harness/simple/jre`.
With java 8, the `java-8-openjdk` folder contained a `jre` folder. To embedd the jvm it was enough to include this folder in the shipped application and load the jvm from there. I don't know what contents of `java-14-openjdk` are needed for this to work. When i copied everything it worked.
Maybe @raniejade knows what contents are needed for the embedding to work?